[VGNWB-254] Move template search from toolbar to tab
