+++ src/components/annotator/ToolBar.vue
@@ -4,19 +4,6 @@
 		v-spacer
 		v-toolbar-title.text-capitalize {{ text.title }}
 		v-spacer
-		v-menu(bottom left offset-y)
-			template(v-slot:activator="{ on }")
-				v-text-field(hide-details v-model="query" filled rounded single-line label="Search for Relation Templates" dense)
-					template(v-slot:append)
-						v-fade-transition(leave-absolute)
-							v-icon(v-on="on" @click="searchRelationTemplates()") search
-			v-list(two-line)
-				template(v-for="(template, i) in templates" )
-					v-list-item(:key="i" @click="showTemplate(template)")
-						v-list-item-content(class="template-list")
-							v-list-item-title(v-html="template.name")
-							v-list-item-subtitle(v-html="template.description")
-					v-divider(v-if="i + 1 < templates.length" )
 		v-btn(class="mx-2" small color="primary" v-if="!submitted && relationsets && relationsets.length > 0" @click="submitRelationsDialog = true")
 			v-icon(left small) mdi-sync
 			| Submit relations
@@ -41,8 +28,7 @@ import { AxiosResponse } from 'axios';
 import { Component, Prop, Vue } from 'vue-property-decorator';
 
 import AnnotationList from '@/components/relations/AnnotationList.vue';
-import { VForm } from '@/interfaces/GlobalTypes';
-import { RelationSet, RelationTemplate } from '@/interfaces/RelationTypes';
+import { RelationSet } from '@/interfaces/RelationTypes';
 
 @Component({
 	name: 'ToolBar',
@@ -54,10 +40,7 @@ export default class ToolBar extends Vue {
 	@Prop() private text!: object;
 	@Prop() private relationsets!: RelationSet[];
 
-	private query: string = '';
 	private loading: boolean = false;
-	private showResults: boolean = false;
-	private templates: RelationTemplate[] = [];
 
 	private submitted: boolean = false;
 	private submitting: boolean = false;
@@ -72,38 +55,6 @@ export default class ToolBar extends Vue {
 
 	private showLists() {
 		this.$store.commit('toggleLists');
-	}
-
-	private searchRelationTemplates(): void {
-		this.loading = true;
-		let all = true;
-		if (this.query !== '') {
-			all = false;
-		}
-		Vue.$axios.get('/relationtemplate', {
-			params: {
-				format: 'json',
-				all,
-				search: this.query,
-			},
-		}).then((result) => {
-			this.loading = false;
-			this.showResults = true;
-			this.templates = result.data;
-		}).catch((error) => {
-			// TODO: deal with errors
-		});
-	}
-
-	private showTemplate(template: RelationTemplate): void {
-		this.$store.commit('setAnnotatorCurrentTab', 'tab-3');
-		this.$store.commit('setAnnotatorTemplate', template);
-
-		let fieldAnnotations: any[] = [];
-		if (template.fields) {
-			fieldAnnotations = template.fields.map((i) => null);
-		}
-		this.$store.commit('setSelectedFieldAnnotations', fieldAnnotations);
 	}
 
 	private submitRelations() {
